+++ packages/analytics-core/src/index.ts
@@ -154,11 +154,13 @@ export { NodeConfig, NodeOptions } from './types/config/node-config';
 // React Native
 export {
   ReactNativeConfig,
+  ReactNativeConfigAutocaptureBeta,
+  ReactNativeAutocaptureOptions,
   ReactNativeTrackingOptions,
   ReactNativeOptions,
   ReactNativeAttributionOptions,
 } from './types/config/react-native-config';
-export { ReactNativeClient } from './types/client/react-native-client';
+export { ReactNativeClient, NavigationState } from './types/client/react-native-client';
 
 export { Observable, asyncMap, merge, multicast, Unsubscribable } from './utils/observable';