+++ packages/analytics-react-native-test/.harness/manifest.js
@@ -1,0 +1,4 @@
+global.RN_HARNESS = { 
+    appRegistryComponentName: 'app',
+    disableViewFlattening: false,
+  };