+++ packages/analytics-core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.53.1](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-core@2.53.0...@amplitude/analytics-core@2.53.1) (2026-07-17)
+
+**Note:** Version bump only for package @amplitude/analytics-core
+
+
+
+
+
 # [2.53.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-core@2.52.2...@amplitude/analytics-core@2.53.0) (2026-07-14)